requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.1.0
bleach==3.1.5
cachetools==4.1.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.19
decorator==4.4.2
defusedxml==0.6.0
document==1.0
doqu==0.28.2
entrypoints==0.3
filterpy==1.4.5
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imageio==2.8.0
imutils==0.5.3
iniconfig==1.1.1
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.33.0
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.7.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numba==0.50.0
numpy==1.18.4
oauthlib==3.1.0
object-detection==0.1
olefile==0.46
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.9
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
pigpio==1.78
Pillow==8.1.0
pluggy==0.13.1
plumbum==1.6.9
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.11.4
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==1.9.6
Pygments==2.6.1
pygtail==0.11.1
pyparsing==2.4.7
pyPS4Controller==1.2.1
pyrsistent==0.16.0
pyserial==3.4
pytest==6.2.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
rpi-hardware-pwm==0.0.2
rpyc==4.1.5
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
sh==1.13.1
sip==4.19.13
six==1.15.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.12.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.1
torch==1.5.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
ui==0.1.4
urllib3==1.25.9
utils==1.0.1
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
